+++ packages/package-json-lint-config-terra/package-json-lint.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   rules: {
+    'require-ie10-compatible-dependency-versions': { severityType: 'warn' },
     'require-no-hard-coded-dependency-versions': { severityType: 'error' },
     'require-no-terra-base-peer-dependency-versions': { severityType: 'warn' },
     'require-theme-context-versions': { severityType: 'warn' },