+++ packages/package-json-lint/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added new rule `require-ie10-compatible-dependency-versions`.
+
 ## 1.2.1 - (September 30, 2021)
 
 * Changed